address roughly 85% of comments
