+++ cot-cli/Cargo.toml
@@ -21,8 +21,9 @@ workspace = true
 
 [dependencies]
 anyhow.workspace = true
-assert_cmd = {workspace = true, optional = true}
+assert_cmd = { workspace = true, optional = true }
 cargo_toml.workspace = true
+cargo_metadata.workspace = true
 chrono.workspace = true
 clap = { workspace = true, features = ["derive", "env", "wrap_help", "string"] }
 clap_complete.workspace = true
@@ -45,7 +46,7 @@ tracing-subscriber = { workspace = true, features = ["env-filter"] }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true}
 wait-timeout = { workspace = true }
-tempfile = {workspace = true, optional = true}
+tempfile = { workspace = true, optional = true }
 
 [dev-dependencies]
 cot-cli = { path = ".", features = ["test_utils"] }
